Fix offset calculation when clipping to zoom 0.
We don't generate metatiles with negative zoom levels, so when we get a request which would cause us to request a negative zoom metatile, we adjust the offset to keep the metatile zoom at zero. Previously, we were just clamping the metatile and offset to `0/0/0`, but as we make larger metatiles this no longer works for low zooms.