Configurable layer parameter for S3 keys.
This allows us to match a code change made in https://github.com/tilezen/tilequeue/pull/344, where the layer (and leading slash) are not longer part of the hashed string to make the S3 key.

Added tests to make sure we're conforming to the new way of calculating hashes and still backwards compatible with the old way.